# frontend/src/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent-eBPF MCP Gateway",
    "description": "Production-Grade Security Gateway with eBPF Integration",
    "version": "2.0.0-ULTRA"
  },
  "paths": {
...
    "/api/v1/intent/lease": {
      "post": {
        "tags": [
          "Intent-to-Execution"
        ],
        "summary": "Grant Intent Lease",
        "description": "Evaluates and provisions an in-kernel eBPF Intent Lease for an autonomous AI agent process.",
        "operationId": "grant_intent_lease_api_v1_intent_lease_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntentLeaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntentLeaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/containment/isolate": {
      "post": {
        "tags": [
          "Incident Containment"
        ],
        "summary": "Isolate Process",
        "description": "Triggers automated quarantine on target process (cgroupv2 SIGSTOP) and generates forensics snapshot.",
        "operationId": "isolate_process_api_v1_containment_isolate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IsolateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/containment/release": {
      "post": {
        "tags": [
          "Incident Containment"
        ],
        "summary": "Release Process",
        "description": "Releases an isolated process back into normal execution (SIGCONT).",
        "operationId": "release_process_api_v1_containment_release_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/mesh/register": {
      "post": {
        "tags": [
          "Distributed Telemetry Mesh"
        ],
        "summary": "Register Node",
        "description": "Registers an edge telemetry agent into the global control plane mesh.",
        "operationId": "register_node_api_v1_mesh_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EdgeNode"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EdgeNode"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mesh/heartbeat": {
      "post": {
        "tags": [
          "Distributed Telemetry Mesh"
        ],
        "summary": "Node Heartbeat",
        "description": "Processes liveness heartbeats and dynamic telemetry statistics from edge nodes.",
        "operationId": "node_heartbeat_api_v1_mesh_heartbeat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EdgeNode"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/billing/create-checkout-session": {
      "post": {
        "tags": [
          "Stripe Billing"
        ],
        "summary": "Create Stripe Checkout Session",
        "description": "Creates an authenticated Stripe Checkout Session containing metadata for automated quota boosts.",
        "operationId": "create_stripe_checkout_session_api_v1_billing_create_checkout_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutSessionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckoutSessionCreateRequest": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id",
            "description": "Customer Tenant ID receiving quota boost"
          },
          "plan_tier": {
            "type": "string",
            "title": "Plan Tier",
            "description": "Subscription plan tier (pro, enterprise)",
            "default": "enterprise"
          },
          "quota_boost": {
            "type": "integer",
            "title": "Quota Boost",
            "description": "Additional eBPF telemetry inspections credited",
            "default": 250000
          },
          "success_url": {
            "type": "string",
            "title": "Success Url",
            "default": "https://ksec.space/dashboard/billing?success=true"
          },
          "cancel_url": {
            "type": "string",
            "title": "Cancel Url",
            "default": "https://ksec.space/dashboard/billing?canceled=true"
          }
        },
        "type": "object",
        "required": [
          "tenant_id"
        ],
        "title": "CheckoutSessionCreateRequest"
      },
...
      "EdgeNode": {
        "properties": {
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "hostname": {
            "type": "string",
            "title": "Hostname"
          },
          "region": {
            "type": "string",
            "title": "Region",
            "default": "us-east-1"
          },
          "ip_address": {
            "type": "string",
            "title": "Ip Address"
          },
          "ebpf_probes_loaded": {
            "type": "integer",
            "title": "Ebpf Probes Loaded",
            "default": 1
          },
          "active_kprobes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Active Kprobes",
            "default": [
              "tcp_v4_connect"
            ]
          },
          "cpu_usage_pct": {
            "type": "number",
            "title": "Cpu Usage Pct",
            "default": 0.0
          },
          "memory_mb": {
            "type": "number",
            "title": "Memory Mb",
            "default": 0.0
          },
          "last_heartbeat_timestamp": {
            "type": "number",
            "title": "Last Heartbeat Timestamp"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "ONLINE"
          }
        },
        "type": "object",
        "required": [
          "node_id",
          "hostname",
          "ip_address"
        ],
        "title": "EdgeNode"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HeartbeatPayload": {
        "properties": {
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "cpu_usage_pct": {
            "type": "number",
            "title": "Cpu Usage Pct",
            "default": 0.0
          },
          "memory_mb": {
            "type": "number",
            "title": "Memory Mb",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "node_id"
        ],
        "title": "HeartbeatPayload"
      },
...
      "IncidentReport": {
        "properties": {
          "incident_id": {
            "type": "string",
            "title": "Incident Id"
          },
          "pid": {
            "type": "integer",
            "title": "Pid"
          },
          "comm": {
            "type": "string",
            "title": "Comm"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "timestamp_ns": {
            "type": "integer",
            "title": "Timestamp Ns"
          },
          "action_taken": {
            "type": "string",
            "title": "Action Taken"
          },
          "forensics_snapshot": {
            "additionalProperties": true,
            "type": "object",
            "title": "Forensics Snapshot"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "CONTAINED"
          }
        },
        "type": "object",
        "required": [
          "incident_id",
          "pid",
          "comm",
          "reason",
          "timestamp_ns",
          "action_taken",
          "forensics_snapshot"
        ],
        "title": "IncidentReport"
      },
      "IntentLeaseRequest": {
        "properties": {
          "intent_id": {
            "type": "string",
            "title": "Intent Id",
            "description": "Unique Intent Identifier (e.g. intent-net-01)"
          },
          "pid": {
            "type": "integer",
            "title": "Pid",
            "description": "Target Process ID running under agent execution context"
          },
          "action": {
            "type": "string",
            "enum": [
              "ALLOW",
              "DENY",
              "allow",
              "deny"
            ],
            "title": "Action",
            "description": "Policy action verdict"
          }
        },
        "type": "object",
        "required": [
          "intent_id",
          "pid",
          "action"
        ],
        "title": "IntentLeaseRequest"
      },
      "IntentLeaseResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Execution status",
            "default": "success"
          },
          "intent_id": {
            "type": "string",
            "title": "Intent Id",
            "description": "Granted intent identifier"
          },
          "pid": {
            "type": "integer",
            "title": "Pid",
            "description": "Target Process ID"
          },
          "enforced_in_kernel": {
            "type": "boolean",
            "title": "Enforced In Kernel",
            "description": "Whether policy was synchronized to eBPF map",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "intent_id",
          "pid"
        ],
        "title": "IntentLeaseResponse"
      },
      "IsolateRequest": {
        "properties": {
          "pid": {
            "type": "integer",
            "title": "Pid",
            "description": "Target process PID to immediately freeze and isolate"
          },
          "comm": {
            "type": "string",
            "title": "Comm",
            "description": "Process name / binary executable",
            "default": "unknown"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "description": "Containment rationale",
            "default": "Manual security quarantine triggered from dashboard"
          }
        },
        "type": "object",
        "required": [
          "pid"
        ],
        "title": "IsolateRequest"
      },
...
      "ReleaseRequest": {
        "properties": {
          "incident_id": {
            "type": "string",
            "title": "Incident Id",
            "description": "Target incident ID to release and unfreeze"
          }
        },
        "type": "object",
        "required": [
          "incident_id"
        ],
        "title": "ReleaseRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}